# docs/devicemanager_swagger.yaml
basePath: /
definitions:
...
  models.Location:
    properties:
      description:
        type: string
      device_group_ids:
        items:
          type: string
        type: array
      device_ids:
        items:
          type: string
        type: array
      id:
        type: string
      image:
        type: string
      name:
        type: string
    type: object
  models.Protocol:
    properties:
      constraints:
        items:
          type: string
        type: array
      handler:
        type: string
      id:
        type: string
      name:
        type: string
      protocol_segments:
        items:
          $ref: '#/definitions/models.ProtocolSegment'
        type: array
    type: object
  models.ProtocolSegment:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Manager API
  version: "0.1"
paths:
...
  /locations:
    post:
      description: create location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: create location
      tags:
      - create
      - locations
  /locations/{id}:
    delete:
      description: delete location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete location
      tags:
      - delete
      - locations
    get:
      description: get location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get location
      tags:
      - get
      - locations
    put:
      description: set location
      parameters:
      - description: Location Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set location
      tags:
      - set
      - locations
  /protocols:
    post:
      description: get protocol
      parameters:
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Protocol'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Protocol'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get protocol
      tags:
      - get
      - protocols
  /protocols/{id}:
    delete:
      description: delete protocol
      parameters:
      - description: Protocol Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: delete protocol
      tags:
      - delete
      - protocols
    get:
      description: get protocol
      parameters:
      - description: Protocol Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Protocol'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get protocol
      tags:
      - get
      - protocols
    put:
      description: set protocol
      parameters:
      - description: Protocol Id
        in: path
        name: id
        required: true
        type: string
      - description: wait for done message in kafka before responding
        in: query
        name: wait
        type: boolean
      - description: element
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Protocol'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Protocol'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: set protocol
      tags:
      - set
      - protocols
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"